Extract transformer
