+++ test/unit/transformers.spec.js
@@ -1,0 +1,19 @@
+import { expect } from 'chai';
+import { transformSummaryStats } from '../../src/ui-data/transformers';
+
+describe('Transformers', () => {
+  it('transforms summary stats data into the UI shape', () => {
+    const expected = [
+      {id: 1, label: "Wins", text: "3"},
+      {id: 2, label: "Losses", text: "2"},
+      {id: 3, label: "Ratio", text: "60%"}
+    ];
+    const input = {
+      wins: 3,
+      losses: 2,
+      ratio: 0.6 * 100 + '%'
+    };
+
+    expect(expected).to.deep.equal(transformSummaryStats(input));
+  });
+});
